refactor: function which configures tls with pkcs12 file now a helper function in util module
