+++ zeebe/atomix/cluster/src/main/java/io/atomix/cluster/messaging/impl/NettyMessagingService.java
@@ -29,6 +29,7 @@ import io.atomix.cluster.messaging.MessagingService;
 import io.atomix.utils.concurrent.OrderedFuture;
 import io.atomix.utils.net.Address;
 import io.camunda.zeebe.util.StringUtil;
+import io.camunda.zeebe.util.TlsConfigUtil;
 import io.camunda.zeebe.util.VisibleForTesting;
 import io.netty.bootstrap.Bootstrap;
 import io.netty.bootstrap.ServerBootstrap;
@@ -68,19 +69,10 @@ import io.netty.resolver.dns.DnsNameResolverBuilder;
 import io.netty.resolver.dns.LoggingDnsQueryLifeCycleObserverFactory;
 import io.netty.util.concurrent.DefaultThreadFactory;
 import io.netty.util.concurrent.Future;
-import java.io.File;
-import java.io.FileInputStream;
 import java.net.ConnectException;
 import java.net.InetSocketAddress;
-import java.security.KeyStore;
-import java.security.KeyStoreException;
-import java.security.NoSuchAlgorithmException;
-import java.security.PrivateKey;
-import java.security.UnrecoverableKeyException;
-import java.security.cert.X509Certificate;
 import java.time.Duration;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.List;
@@ -486,7 +478,7 @@ public final class NettyMessagingService implements ManagedMessagingService {
 
       if (config.getKeyStore() != null) {
         sslContextBuilder.trustManager(
-            getCertificateChain(config.getKeyStore(), config.getKeyStorePassword()));
+            TlsConfigUtil.getCertificateChain(config.getKeyStore(), config.getKeyStorePassword()));
       } else {
         sslContextBuilder.trustManager(config.getCertificateChain());
       }
@@ -505,9 +497,10 @@ public final class NettyMessagingService implements ManagedMessagingService {
       final SslContextBuilder sslContextBuilder;
 
       if (config.getKeyStore() != null) {
-        final var privateKey = getPrivateKey(config.getKeyStore(), config.getKeyStorePassword());
+        final var privateKey =
+            TlsConfigUtil.getPrivateKey(config.getKeyStore(), config.getKeyStorePassword());
         final var certChain =
-            getCertificateChain(config.getKeyStore(), config.getKeyStorePassword());
+            TlsConfigUtil.getCertificateChain(config.getKeyStore(), config.getKeyStorePassword());
 
         sslContextBuilder = SslContextBuilder.forServer(privateKey, certChain);
       } else {
@@ -530,41 +523,6 @@ public final class NettyMessagingService implements ManagedMessagingService {
     } else {
       initNioTransport();
     }
-  }
-
-  private X509Certificate[] getCertificateChain(final File keyStoreFile, final String password)
-      throws KeyStoreException {
-    final var keyStore = getKeyStore(keyStoreFile, password);
-
-    final String alias = keyStore.aliases().nextElement();
-    return Arrays.stream(keyStore.getCertificateChain(alias))
-        .map(X509Certificate.class::cast)
-        .toArray(X509Certificate[]::new);
-  }
-
-  private PrivateKey getPrivateKey(final File keyStoreFile, final String password)
-      throws KeyStoreException, NoSuchAlgorithmException, UnrecoverableKeyException {
-
-    final var keyStore = getKeyStore(keyStoreFile, password);
-
-    final String alias = keyStore.aliases().nextElement();
-    return (PrivateKey) keyStore.getKey(alias, password.toCharArray());
-  }
-
-  private KeyStore getKeyStore(final File keyStoreFile, final String password)
-      throws KeyStoreException {
-    final var keyStore = KeyStore.getInstance("PKCS12");
-    try {
-      keyStore.load(new FileInputStream(keyStoreFile), password.toCharArray());
-    } catch (final Exception e) {
-      throw new IllegalStateException(
-          String.format(
-              "Keystore failed to load file: %s, please ensure it is a valid PKCS12 keystore",
-              keyStoreFile.toPath()),
-          e);
-    }
-
-    return keyStore;
   }
 
   private void initEpollTransport() {
